+++ ESI.NET/EsiResponse.cs
@@ -1,7 +1,5 @@
 ﻿using Newtonsoft.Json;
 using System;
-using System.Collections.Generic;
-using System.Collections.Immutable;
 using System.Linq;
 using System.Net;
 using System.Net.Http;
@@ -66,7 +64,7 @@ namespace ESI.NET
                     ErrorLimitReset = int.Parse(response.Headers.GetValues("X-Esi-Error-Limit-Reset").First());
 
                 if (response.StatusCode == HttpStatusCode.NoContent)
-                    Message = _noContentMessage.TryGetValue(path, out var noContent) ? noContent : "No Content";
+                    Message = "No Content";
                 else if (response.StatusCode == HttpStatusCode.OK ||
                          response.StatusCode == HttpStatusCode.Created)
                 {
@@ -105,45 +103,5 @@ namespace ESI.NET
         public string Message { get; set; }
         public T Data { get; set; }
         public Exception Exception { get; set; }
-
-        private static readonly ImmutableDictionary<string, string> _noContentMessage = new Dictionary<string, string>()
-        {
-            //Calendar
-            {"PUT|/characters/{character_id}/calendar/{event_id}/", "Event updated"},
-
-            //Contacts
-            {"PUT|/characters/{character_id}/contacts/", "Contacts updated"},
-            {"DELETE|/characters/{character_id}/contacts/", "Contacts deleted"},
-
-            //Corporations
-            {"PUT|/corporations/{corporation_id}/structures/{structure_id}/", "Structure vulnerability window updated"},
-
-            //Fittings
-            {"DELETE|/characters/{character_id}/fittings/{fitting_id}/", "Fitting deleted"},
-
-            //Fleets
-            {"PUT|/fleets/{fleet_id}/", "Fleet updated"},
-            {"POST|/fleets/{fleet_id}/members/", "Fleet invitation sent"},
-            {"DELETE|/fleets/{fleet_id}/members/{member_id}/", "Fleet member kicked"},
-            {"PUT|/fleets/{fleet_id}/members/{member_id}/", "Fleet invitation sent"},
-            {"DELETE|/fleets/{fleet_id}/wings/{wing_id}/", "Wing deleted"},
-            {"PUT|/fleets/{fleet_id}/wings/{wing_id}/", "Wing renamed"},
-            {"DELETE|/fleets/{fleet_id}/squads/{squad_id}/", "Squad deleted"},
-            {"PUT|/fleets/{fleet_id}/squads/{squad_id}/", "Squad renamed"},
-
-            //Mail
-            {"POST|/characters/{character_id}/mail/", "Mail created"},
-            {"POST|/characters/{character_id}/mail/labels/", "Label created"},
-            {"DELETE|/characters/{character_id}/mail/labels/{label_id}/", "Label deleted"},
-            {"PUT|/characters/{character_id}/mail/{mail_id}/", "Mail updated"},
-            {"DELETE|/characters/{character_id}/mail/{mail_id}/", "Mail deleted"},
-
-            //User Interface
-            {"POST|/ui/openwindow/marketdetails/", "Open window request received"},
-            {"POST|/ui/openwindow/contract/", "Open window request received"},
-            {"POST|/ui/openwindow/information/", "Open window request received"},
-            {"POST|/ui/autopilot/waypoint/", "Open window request received"},
-            {"POST|/ui/openwindow/newmail/", "Open window request received"}
-        }.ToImmutableDictionary();
     }
 }